added rollback to randomwalk

+++ src/main/java/ed/mse/graph/Pair.java
@@ -5,8 +5,8 @@ import ed.mse.graph.Graph.Node;
 public class Pair {
 	Node node; 
 	Node via; 
-	int distance; 
-	public Pair(Node node, Node via, int distance) {
+	double distance; 
+	public Pair(Node node, Node via, double distance) {
 		this.node = node;
 		this.via = via; 
 		this.distance = distance;
